Deal with pseudo-headers.
Context: We made `IncomingRequest.headers` be an `HttpHeaders` object, but
its superclass, which is the built-in class `Headers` does not accept
header names in the pseudo-header form (`TypeError`). Long-term we should
probably make `HttpHeaders` _not_ be a subclass of `Headers`, but in the
meantime what we're doing here is pretty reasonable.