Make visiting happen as a single "thread."
Previously, it was possible for a visit to have multiple pending promises
for parts of the visit, which could end up confusing the detection of shared
references, including circular references. Now, there is only ever one
pending promise for a visit result.

This also means that the documented contract of visiting "depth first" is
now actually honored.